Move author GUI to ui.tk
